feat(web): feature based folded structure
