+++ apps/web/src/lib/errors.ts
