+++ apps/web/src/middleware.ts
@@ -1,4 +1,4 @@
-import { sessionOptions } from '@/lib/api/auth/session.config';
+import { sessionOptions } from '@/features/auth/config/session.config';
 import { SessionData } from '@repo/types';
 import { getIronSession } from 'iron-session';
 import { NextRequest, NextResponse } from 'next/server';
